pyproject.toml
[project]
name = "bigq-kaggle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery>=3.36.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "google-genai>=1.38.0",
    "gradio>=5.44.1",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "google-cloud-aiplatform>=1.60.0",
    "pytest>=8.4.2",
    "pdfplumber>=0.11.7",
    "google-cloud-storage>=2.18.2",
    "bigframes>=1.31.0",
    "google-cloud-vision>=3.10.2",
    "bigframes>=2.21.0",
]